+++ src/Razor/src/Microsoft.AspNetCore.Razor.Language/RazorCodeDocumentExtensions.cs
@@ -115,6 +115,25 @@ namespace Microsoft.AspNetCore.Razor.Language
             document.Items[typeof(DocumentIntermediateNode)] = documentNode;
         }
 
+        internal static RazorHtmlDocument GetHtmlDocument(this RazorCodeDocument document)
+        {
+            if (document == null)
+            {
+                throw new ArgumentNullException(nameof(document));
+            }
+
+            var razorHtmlObj = document.Items[typeof(RazorHtmlDocument)];
+            if (razorHtmlObj == null)
+            {
+                var razorHtmlDocument = RazorHtmlWriter.GetHtmlDocument(document);
+                document.Items[typeof(RazorHtmlDocument)] = razorHtmlDocument;
+
+                return razorHtmlDocument;
+            }
+
+            return (RazorHtmlDocument)razorHtmlObj;
+        }
+
         public static RazorCSharpDocument GetCSharpDocument(this RazorCodeDocument document)
         {
             if (document == null)
@@ -133,26 +152,6 @@ namespace Microsoft.AspNetCore.Razor.Language
             }
 
             document.Items[typeof(RazorCSharpDocument)] = csharp;
-        }
-
-        internal static RazorHtmlDocument GetHtmlDocument(this RazorCodeDocument document)
-        {
-            if (document == null)
-            {
-                throw new ArgumentNullException(nameof(document));
-            }
-
-            return (RazorHtmlDocument)document.Items[typeof(RazorHtmlDocument)];
-        }
-
-        internal static void SetHtmlDocument(this RazorCodeDocument document, RazorHtmlDocument html)
-        {
-            if (document == null)
-            {
-                throw new ArgumentNullException(nameof(document));
-            }
-
-            document.Items[typeof(RazorHtmlDocument)] = html;
         }
 
         public static RazorParserOptions GetParserOptions(this RazorCodeDocument document)
