+++ src/Razor/src/Microsoft.AspNetCore.Razor.Language/RazorHtmlDocument.cs
@@ -2,7 +2,6 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Collections.Generic;
 
 namespace Microsoft.AspNetCore.Razor.Language
 {
@@ -12,7 +11,7 @@ namespace Microsoft.AspNetCore.Razor.Language
 
         public abstract RazorCodeGenerationOptions Options { get; }
 
-        public static RazorHtmlDocument Create(string generatedHtml, RazorCodeGenerationOptions options, IEnumerable<RazorDiagnostic> diagnostics)
+        public static RazorHtmlDocument Create(string generatedHtml, RazorCodeGenerationOptions options)
         {
             if (generatedHtml == null)
             {
